Deleting messages and changing message properties was not possible if correlation key reused between messages (#125)
* avoid error when message ref doesn't exist on correlation property (rare, but ran into it).  When deleting an object with a message, be aware that you might be deeply nested in other things.

* MatchingConditionArray, TaskEventMessageProvider.js - fixing the ids of div elements to make them easier to find in tests.

MessageHelpers - And added a deleteMessage function that will remove a message AND it's correlation properties. (The reason I came here)

MessageHelpers, MessageInterceptor.js - fixing the misspelled name of syncCorrelationProperties.

MessageCorrSpec tests that when you update a message through a message editor, it will correctly update the correlation properties, and not leave invalid properties lying around.

MessageSelect - will now remove the old message definition (including correlation properties), and replace it with the new message definition when a message is returned from the message editor.  As we do this when we receive a new message,there is no need for the cleanupOldMessage function when selecting a different message.

MessageSpec - a hell of a time with this.  "spiffExtensionOptions" is a constant that I could not seem to clear out between differest spec file tests.  I finally resorted to creating a "clearMessages" function that would force the list of messages to get cleared out when needed.

Co-authored-by: burnettk
Co-authored-by: jasquat

* forgot a test file in last commit.

* fixing a minor error.

* Update test/spec/MessagesSpec.js

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* Update test/spec/MessagesSpec.js

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

---------

Co-authored-by: Kevin Burnett <18027+burnettk@users.noreply.github.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>